Text-Extract-Images/requirements.txt
pytesseract==0.3.6
Pillow==9.0.1